Refactored LC call number methods.
